# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "DeadHostCreate": {
        "description": "Payload to create a dead host.",
        "properties": {
          "advanced_config": {
            "default": "",
            "description": "Raw nginx directives injected into the server block",
            "title": "Advanced Config",
            "type": "string"
          },
          "certificate_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Certificate for TLS termination; null serves plain :80",
            "title": "Certificate Id"
          },
          "domain_names": {
            "description": "Domains this host answers for (server_name)",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Domain Names",
            "type": "array"
          },
          "enabled": {
            "default": true,
            "description": "Disabled hosts are excluded from config",
            "title": "Enabled",
            "type": "boolean"
          },
          "hsts_enabled": {
            "default": false,
            "description": "Emit a Strict-Transport-Security header",
            "title": "Hsts Enabled",
            "type": "boolean"
          },
          "hsts_subdomains": {
            "default": false,
            "description": "Include subdomains in HSTS",
            "title": "Hsts Subdomains",
            "type": "boolean"
          },
          "http2_support": {
            "default": false,
            "description": "Enable HTTP/2 on the TLS listener",
            "title": "Http2 Support",
            "type": "boolean"
          },
          "ssl_forced": {
            "default": false,
            "description": "Redirect :80 to HTTPS",
            "title": "Ssl Forced",
            "type": "boolean"
          }
        },
        "required": [
          "domain_names"
        ],
        "title": "DeadHostCreate",
        "type": "object"
      },
      "DeadHostRead": {
        "description": "Public representation of a dead host.",
        "properties": {
          "advanced_config": {
            "default": "",
            "description": "Raw nginx directives injected into the server block",
            "title": "Advanced Config",
            "type": "string"
          },
          "certificate_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Certificate for TLS termination; null serves plain :80",
            "title": "Certificate Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "domain_names": {
            "description": "Domains this host answers for (server_name)",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Domain Names",
            "type": "array"
          },
          "enabled": {
            "default": true,
            "description": "Disabled hosts are excluded from config",
            "title": "Enabled",
            "type": "boolean"
          },
          "hsts_enabled": {
            "default": false,
            "description": "Emit a Strict-Transport-Security header",
            "title": "Hsts Enabled",
            "type": "boolean"
          },
          "hsts_subdomains": {
            "default": false,
            "description": "Include subdomains in HSTS",
            "title": "Hsts Subdomains",
            "type": "boolean"
          },
          "http2_support": {
            "default": false,
            "description": "Enable HTTP/2 on the TLS listener",
            "title": "Http2 Support",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "ssl_forced": {
            "default": false,
            "description": "Redirect :80 to HTTPS",
            "title": "Ssl Forced",
            "type": "boolean"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "domain_names",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "DeadHostRead",
        "type": "object"
      },
      "DeadHostUpdate": {
        "description": "Partial update of a dead host; every field is optional.",
        "properties": {
          "advanced_config": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advanced Config"
          },
          "certificate_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Certificate Id"
          },
          "domain_names": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "minItems": 1,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Names"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "hsts_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hsts Enabled"
          },
          "hsts_subdomains": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hsts Subdomains"
          },
          "http2_support": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Http2 Support"
          },
          "ssl_forced": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssl Forced"
          }
        },
        "title": "DeadHostUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NginxConfigFile": {
        "description": "One rendered managed ``.conf`` file.",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "content"
        ],
        "title": "NginxConfigFile",
        "type": "object"
      },
      "NginxConfigPreview": {
        "description": "The full config the engine *would* write for the current DB state.\n\nA read-only render — it does not touch disk or reload nginx, so the\nfrontend can show operators the generated output before/without applying.",
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/NginxConfigFile"
            },
            "title": "Files",
            "type": "array"
          }
        },
        "title": "NginxConfigPreview",
        "type": "object"
      },
...
      "TaskEnqueued": {
        "description": "Returned when a task is accepted onto the queue.",
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "status"
        ],
        "title": "TaskEnqueued",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "MegooPM",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/dead-hosts": {
      "get": {
        "description": "List all dead hosts. Admin-only.",
        "operationId": "list_dead_hosts_api_v1_dead_hosts_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DeadHostRead"
                  },
                  "title": "Response List Dead Hosts Api V1 Dead Hosts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Dead Hosts",
        "tags": [
          "dead-hosts"
        ]
      },
      "post": {
        "description": "Create a dead host. Admin-only.",
        "operationId": "create_dead_host_api_v1_dead_hosts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeadHostCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeadHostRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Dead Host",
        "tags": [
          "dead-hosts"
        ]
      }
    },
    "/api/v1/dead-hosts/{host_id}": {
      "delete": {
        "description": "Delete a dead host. Admin-only.",
        "operationId": "delete_dead_host_api_v1_dead_hosts__host_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "host_id",
            "required": true,
            "schema": {
              "title": "Host Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Dead Host",
        "tags": [
          "dead-hosts"
        ]
      },
      "get": {
        "description": "Fetch a single dead host. Admin-only.",
        "operationId": "get_dead_host_api_v1_dead_hosts__host_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "host_id",
            "required": true,
            "schema": {
              "title": "Host Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeadHostRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Dead Host",
        "tags": [
          "dead-hosts"
        ]
      },
      "patch": {
        "description": "Update a dead host. Admin-only.",
        "operationId": "update_dead_host_api_v1_dead_hosts__host_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "host_id",
            "required": true,
            "schema": {
              "title": "Host Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeadHostUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeadHostRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Dead Host",
        "tags": [
          "dead-hosts"
        ]
      }
    },
    "/api/v1/nginx/preview": {
      "get": {
        "description": "Render the config for current DB state without writing or reloading.",
        "operationId": "preview_nginx_config_api_v1_nginx_preview_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NginxConfigPreview"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Preview Nginx Config",
        "tags": [
          "nginx"
        ]
      }
    },
    "/api/v1/nginx/reload": {
      "post": {
        "description": "Enqueue a config regeneration + nginx reload; returns a task id to poll.",
        "operationId": "reload_nginx_api_v1_nginx_reload_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskEnqueued"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Reload Nginx",
        "tags": [
          "nginx"
        ]
      }
    },
...
  }
}